requirements.txt
requests
pandas
numpy
matplotlib
plotly
transformers
torch
tensorflow
flask
langchain
python-dotenv